fix PIE support for CMake >= 3.14
CMake 3.14 requires you to call check_pie_supported() from the
CheckPieSupported module in order for position-independent executables
to be created when CMAKE_POSITION_INDEPENDENT_CODE is true.

This function does a few additional checks to see if extra linker flags
are necessary to get a position-independent executable. It's mostly
only an issue with the GNU toolchain.